move commands out of form-tester into commands
