+++ src/platform/testing/e2e/cypress/support/index.js
@@ -4,6 +4,28 @@ import 'cypress-plugin-tab';
 
 import './commands';
 
+export const APP_SELECTOR = '#react-root';
+export const ARRAY_ITEM_SELECTOR =
+  'div[name^="topOfTable_"] ~ div.va-growable-background';
+export const FIELD_SELECTOR = 'input, select, textarea';
+export const LOADING_SELECTOR = '.loading-indicator';
+
+// Force interactions on elements, skipping the default checks for the
+// "user interactive" state of an element, potentially saving some time.
+// More importantly, this ensures the interaction will target the actual
+// selected element, which overrides the default behavior that simulates
+// how a real user might try to interact with a target element that has moved.
+// https://github.com/cypress-io/cypress/issues/6165
+export const FORCE_OPTION = { force: true };
+
+// Cypress does not officially support typing without delay.
+// See the main support file for more details.
+export const NO_DELAY_OPTION = { delay: 0 };
+
+// Suppress logs for most commands, particularly calls to wrap and get
+// that are mainly there to support more specific operations.
+export const NO_LOG_OPTION = { log: false };
+
 Cypress.on('window:before:load', window => {
   // Workaround to allow Cypress to intercept requests made with the Fetch API.
   // This forces fetch to fall back to the polyfill, which can get intercepted.
